+++ src/networking/types/data_info.rs
@@ -1,6 +1,6 @@
 //! Module defining the `DataInfo` struct, which represents incoming and outgoing packets and bytes.
 
-use crate::chart::types::chart_type::ChartType;
+use crate::networking::types::data_representation::DataRepr;
 use crate::networking::types::traffic_direction::TrafficDirection;
 use crate::report::types::sort_type::SortType;
 use std::cmp::Ordering;
@@ -23,35 +23,24 @@ pub struct DataInfo {
 }
 
 impl DataInfo {
-    pub fn incoming_packets(&self) -> u128 {
-        self.incoming_packets
-    }
-
-    pub fn outgoing_packets(&self) -> u128 {
-        self.outgoing_packets
-    }
-
-    pub fn incoming_bytes(&self) -> u128 {
-        self.incoming_bytes
-    }
-
-    pub fn outgoing_bytes(&self) -> u128 {
-        self.outgoing_bytes
-    }
-
-    pub fn tot_packets(&self) -> u128 {
-        self.incoming_packets + self.outgoing_packets
-    }
-
-    pub fn tot_bytes(&self) -> u128 {
-        self.incoming_bytes + self.outgoing_bytes
-    }
-
-    pub fn tot_data(&self, chart_type: ChartType) -> u128 {
-        match chart_type {
-            ChartType::Packets => self.tot_packets(),
-            ChartType::Bytes => self.tot_bytes(),
+    pub fn incoming_data(&self, data_repr: DataRepr) -> u128 {
+        match data_repr {
+            DataRepr::Packets => self.incoming_packets,
+            DataRepr::Bytes => self.incoming_bytes,
+            DataRepr::Bits => self.incoming_bytes * 8,
         }
+    }
+
+    pub fn outgoing_data(&self, data_repr: DataRepr) -> u128 {
+        match data_repr {
+            DataRepr::Packets => self.outgoing_packets,
+            DataRepr::Bytes => self.outgoing_bytes,
+            DataRepr::Bits => self.outgoing_bytes * 8,
+        }
+    }
+
+    pub fn tot_data(&self, data_repr: DataRepr) -> u128 {
+        self.incoming_data(data_repr) + self.outgoing_data(data_repr)
     }
 
     pub fn add_packet(&mut self, bytes: u128, traffic_direction: TrafficDirection) {
@@ -103,18 +92,11 @@ impl DataInfo {
         self.final_instant = rhs.final_instant;
     }
 
-    pub fn compare(&self, other: &Self, sort_type: SortType, chart_type: ChartType) -> Ordering {
-        match chart_type {
-            ChartType::Packets => match sort_type {
-                SortType::Ascending => self.tot_packets().cmp(&other.tot_packets()),
-                SortType::Descending => other.tot_packets().cmp(&self.tot_packets()),
-                SortType::Neutral => other.final_instant.cmp(&self.final_instant),
-            },
-            ChartType::Bytes => match sort_type {
-                SortType::Ascending => self.tot_bytes().cmp(&other.tot_bytes()),
-                SortType::Descending => other.tot_bytes().cmp(&self.tot_bytes()),
-                SortType::Neutral => other.final_instant.cmp(&self.final_instant),
-            },
+    pub fn compare(&self, other: &Self, sort_type: SortType, data_repr: DataRepr) -> Ordering {
+        match sort_type {
+            SortType::Ascending => self.tot_data(data_repr).cmp(&other.tot_data(data_repr)),
+            SortType::Descending => other.tot_data(data_repr).cmp(&self.tot_data(data_repr)),
+            SortType::Neutral => other.final_instant.cmp(&self.final_instant),
         }
     }
 
@@ -144,5 +126,117 @@ impl Default for DataInfo {
             outgoing_bytes: 0,
             final_instant: Instant::now(),
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::networking::types::traffic_direction::TrafficDirection;
+
+    #[test]
+    fn test_data_info() {
+        // in_packets: 0, out_packets: 0, in_bytes: 0, out_bytes: 0
+        let mut data_info_1 = DataInfo::new_with_first_packet(123, TrafficDirection::Incoming);
+        // 1, 0, 123, 0
+        data_info_1.add_packet(100, TrafficDirection::Incoming);
+        // 2, 0, 223, 0
+        data_info_1.add_packet(200, TrafficDirection::Outgoing);
+        // 2, 1, 223, 200
+        data_info_1.add_packets(11, 1200, TrafficDirection::Outgoing);
+        // 2, 12, 223, 1400
+        data_info_1.add_packets(5, 500, TrafficDirection::Incoming);
+        // 7, 12, 723, 1400
+
+        assert_eq!(data_info_1.incoming_packets, 7);
+        assert_eq!(data_info_1.outgoing_packets, 12);
+        assert_eq!(data_info_1.incoming_bytes, 723);
+        assert_eq!(data_info_1.outgoing_bytes, 1400);
+
+        assert_eq!(data_info_1.tot_data(DataRepr::Packets), 19);
+        assert_eq!(data_info_1.tot_data(DataRepr::Bytes), 2123);
+        assert_eq!(data_info_1.tot_data(DataRepr::Bits), 16984);
+
+        assert_eq!(data_info_1.incoming_data(DataRepr::Packets), 7);
+        assert_eq!(data_info_1.incoming_data(DataRepr::Bytes), 723);
+        assert_eq!(data_info_1.incoming_data(DataRepr::Bits), 5784);
+
+        assert_eq!(data_info_1.outgoing_data(DataRepr::Packets), 12);
+        assert_eq!(data_info_1.outgoing_data(DataRepr::Bytes), 1400);
+        assert_eq!(data_info_1.outgoing_data(DataRepr::Bits), 11200);
+
+        let mut data_info_2 = DataInfo::new_with_first_packet(100, TrafficDirection::Outgoing);
+        // 0, 1, 0, 100
+        data_info_2.add_packets(19, 300, TrafficDirection::Outgoing);
+        // 0, 20, 0, 400
+
+        assert_eq!(data_info_2.incoming_packets, 0);
+        assert_eq!(data_info_2.outgoing_packets, 20);
+        assert_eq!(data_info_2.incoming_bytes, 0);
+        assert_eq!(data_info_2.outgoing_bytes, 400);
+
+        assert_eq!(data_info_2.tot_data(DataRepr::Packets), 20);
+        assert_eq!(data_info_2.tot_data(DataRepr::Bytes), 400);
+        assert_eq!(data_info_2.tot_data(DataRepr::Bits), 3200);
+
+        assert_eq!(data_info_2.incoming_data(DataRepr::Packets), 0);
+        assert_eq!(data_info_2.incoming_data(DataRepr::Bytes), 0);
+        assert_eq!(data_info_2.incoming_data(DataRepr::Bits), 0);
+
+        assert_eq!(data_info_2.outgoing_data(DataRepr::Packets), 20);
+        assert_eq!(data_info_2.outgoing_data(DataRepr::Bytes), 400);
+        assert_eq!(data_info_2.outgoing_data(DataRepr::Bits), 3200);
+
+        // compare data_info_1 and data_info_2
+
+        assert_eq!(
+            data_info_1.compare(&data_info_2, SortType::Ascending, DataRepr::Packets),
+            Ordering::Less
+        );
+        assert_eq!(
+            data_info_1.compare(&data_info_2, SortType::Descending, DataRepr::Packets),
+            Ordering::Greater
+        );
+        assert_eq!(
+            data_info_1.compare(&data_info_2, SortType::Neutral, DataRepr::Packets),
+            Ordering::Greater
+        );
+
+        assert_eq!(
+            data_info_1.compare(&data_info_2, SortType::Ascending, DataRepr::Bytes),
+            Ordering::Greater
+        );
+        assert_eq!(
+            data_info_1.compare(&data_info_2, SortType::Descending, DataRepr::Bytes),
+            Ordering::Less
+        );
+        assert_eq!(
+            data_info_1.compare(&data_info_2, SortType::Neutral, DataRepr::Bytes),
+            Ordering::Greater
+        );
+
+        assert_eq!(
+            data_info_1.compare(&data_info_2, SortType::Ascending, DataRepr::Bits),
+            Ordering::Greater
+        );
+        assert_eq!(
+            data_info_1.compare(&data_info_2, SortType::Descending, DataRepr::Bits),
+            Ordering::Less
+        );
+        assert_eq!(
+            data_info_1.compare(&data_info_2, SortType::Neutral, DataRepr::Bits),
+            Ordering::Greater
+        );
+
+        // refresh data_info_1 with data_info_2
+        assert!(data_info_1.final_instant < data_info_2.final_instant);
+        data_info_1.refresh(data_info_2);
+
+        // data_info_1 should now contain the sum of both data_info_1 and data_info_2
+        assert_eq!(data_info_1.incoming_packets, 7);
+        assert_eq!(data_info_1.outgoing_packets, 32);
+        assert_eq!(data_info_1.incoming_bytes, 723);
+        assert_eq!(data_info_1.outgoing_bytes, 1800);
+        assert_eq!(data_info_1.final_instant, data_info_2.final_instant);
     }
 }